requirements.txt
nornir==3.0.0
nornir-netmiko==0.1.1
nornir-utils==0.1.0
nornir-jinja2==0.1.0
Faker==4.14.0
PyYAML==5.3.1
ansible==2.9.14